+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+<a name="3.0.4"></a>
+## [3.0.4](https://github.com/bitpay/cordova-plugin-qrscanner/compare/2.6.0...3.0.4) (2020-04-15)
+
+
+### Bug Fixes
+
+* **syntax:** Swift 5 ([692c126](https://github.com/bitpay/cordova-plugin-qrscanner/commit/692c126))
+* **tag:** removes invalid tag UseSwiftLanguageVersion ([3c368fe](https://github.com/bitpay/cordova-plugin-qrscanner/commit/3c368fe))
+
+
+
+<a name="3.0.3"></a>
+## [3.0.3](https://github.com/bitpay/cordova-plugin-qrscanner/compare/2.6.0...3.0.3) (2020-04-15)
+
+
+### Bug Fixes
+
+* **syntax:** Swift 5 ([692c126](https://github.com/bitpay/cordova-plugin-qrscanner/commit/692c126))
+* **tag:** removes invalid tag UseSwiftLanguageVersion ([3c368fe](https://github.com/bitpay/cordova-plugin-qrscanner/commit/3c368fe))
+
+
+
+<a name="3.0.2"></a>
+## [3.0.2](https://github.com/bitpay/cordova-plugin-qrscanner/compare/2.6.0...3.0.2) (2020-04-15)
+
+
+### Bug Fixes
+
+* **syntax:** Swift 5 ([692c126](https://github.com/bitpay/cordova-plugin-qrscanner/commit/692c126))
+* **tag:** removes invalid tag UseSwiftLanguageVersion ([3c368fe](https://github.com/bitpay/cordova-plugin-qrscanner/commit/3c368fe))
+
+
+
 <a name="3.0.1"></a>
 ## [3.0.1](https://github.com/bitpay/cordova-plugin-qrscanner/compare/2.6.0...3.0.1) (2019-03-29)